Fix update quiz date bug. versionCode 9

+++ app/src/main/java/com/zhiyong/tingxie/db/Quiz.java
@@ -21,7 +21,7 @@ public class Quiz {
     }
 
     @Ignore
-    public Quiz(int id, @NonNull int date) {
+    public Quiz(long id, @NonNull int date) {
         this.id = id;
         this.date = date;
     }
